finished logic
